Explicit casts to unsigned int in RemoteMachineComboBox
This silences warnings:
src/RemoteMachineComboBox.cpp:65:36:warning – comparison between signed and unsigned integer expressions [-Wsign-compare]
src/RemoteMachineComboBox.cpp:62:36:warning – comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ src/RemoteMachineComboBox.cpp
@@ -59,10 +59,10 @@ QString RemoteMachineComboBox::getTarget() const
 
 void RemoteMachineComboBox::setRecentMachineCount(unsigned int count)
 {
-    while (mRecentTargets.size() > count)
+    while (static_cast<unsigned int>(mRecentTargets.size()) > count)
         mRecentTargets.removeLast();
 
-    while (mRecentTargets.size() < count)
+    while (static_cast<unsigned int>(mRecentTargets.size()) < count)
         mRecentTargets.append("");
 }
